+++ lib/cffirehol.rb
@@ -897,7 +897,10 @@ module CfFirehol
         fhconf << '# Interfaces'
         fhconf << '#----------------'
         iface_ports.each do |iface, ports|
-            if ifaces.has_key?(iface)
+            if iface.to_s == 'main' and !is_private_iface ifaces[iface]
+                warning("'main' iface is public => packets can arrive on any interface")
+                dev = 'any'
+            elsif ifaces.has_key?(iface)
                 dev = ifaces[iface][:device]
             else
                 # 'any' should be distributed across ifaces